+++ package.json
@@ -37,6 +37,7 @@
     "react-router": "5.1.2",
     "react-router-dom": "5.1.2",
     "react-test-renderer": "16.10.2",
+    "reselect": "4.0.0",
     "ts-jest": "24.2.0",
     "ts-loader": "6.2.1",
     "typescript": "3.7.2",
