+++ src/components/app/app.test.tsx
@@ -63,9 +63,37 @@ describe(`Make snapshot`, () => {
       released: Value.FULL,
       isFavorite: true,
     },
+    filteredFilms: [{
+      id: Value.EMPTY,
+      name: `Keks`,
+      image: {
+        poster: EMPTY_STRING,
+        posterAlt: EMPTY_STRING,
+        preview: EMPTY_STRING,
+        previewAlt: EMPTY_STRING,
+        background: EMPTY_STRING,
+        backgroundAlt: EMPTY_STRING,
+      },
+      backgroundColor: EMPTY_STRING,
+      video: {
+        link: {
+          mp4: EMPTY_STRING,
+          webm: EMPTY_STRING,
+        },
+        poster: EMPTY_STRING,
+      },
+      description: EMPTY_STRING,
+      rating: Value.FULL,
+      scoresCount: Value.FULL,
+      director: `Keks`,
+      starring: [`Me`, `You`, `They`],
+      runTime: Value.EMPTY,
+      genre: EMPTY_STRING,
+      released: Value.FULL,
+      isFavorite: true,
+    }],
     genre: EMPTY_STRING,
     genres: [`ECMA`, `CSS`],
-    iconNames: [`PAUSE`],
     onGenreClick: jest.fn(),
   };
 
@@ -73,9 +101,7 @@ describe(`Make snapshot`, () => {
     const tree = renderer
       .create(
           <BrowserRouter>
-            <App
-              {...initProps}
-            />
+            <App {...initProps}/>
           </BrowserRouter>
       )
       .toJSON();
