+++ src/components/main/main.test.tsx
@@ -12,7 +12,7 @@ jest.mock(`../logo/logo`, () => `Logo`);
 
 describe(`Make snapshot`, () => {
   const initialProps = {
-    films: [{
+    promo: {
       id: Value.EMPTY,
       name: `Keks`,
       image: {
@@ -27,35 +27,8 @@ describe(`Make snapshot`, () => {
       video: {
         link: {
           mp4: EMPTY_STRING,
-          webm: EMPTY_STRING,
         },
         poster: EMPTY_STRING,
-      },
-      description: EMPTY_STRING,
-      rating: Value.FULL,
-      scoresCount: Value.FULL,
-      director: `Keks`,
-      starring: [`Me`, `You`, `They`],
-      runTime: Value.EMPTY,
-      genre: EMPTY_STRING,
-      released: Value.FULL,
-      isFavorite: true,
-    }],
-    promo: {
-      id: Value.EMPTY,
-      name: `Keks`,
-      image: {
-        poster: EMPTY_STRING,
-        posterAlt: `1`,
-        preview: EMPTY_STRING,
-        previewAlt: EMPTY_STRING,
-        background: EMPTY_STRING,
-        backgroundAlt: EMPTY_STRING,
-      },
-      backgroundColor: EMPTY_STRING,
-      video: {
-        link: EMPTY_STRING,
-        preview: EMPTY_STRING,
       },
       description: EMPTY_STRING,
       rating: Value.FULL,
@@ -69,8 +42,8 @@ describe(`Make snapshot`, () => {
     },
     selectedGenre: EMPTY_STRING,
     genres: [`ECMA`, `CSS`],
-    icons: [`PAUSE`],
     onSelectedGenreClick: jest.fn(),
+    onButtonClick: jest.fn(),
   };
 
   it(`Render main screen`, () => {
