+++ src/reducer/name-spaces.js
@@ -1,0 +1,7 @@
+const NameSpaces = {
+  DATA: `DATA`,
+  USER: `USER`,
+};
+
+export default NameSpaces;
+
