+++ src/reducer/user/user.test.js
@@ -1,0 +1,14 @@
+import {initState, ActionType, reducer} from "./user";
+
+describe(`Test actions`, () => {
+
+  it(`Select genre`, () => {
+    expect(reducer(initState, {
+      type: ActionType.SELECT_GENRE,
+      payload: `Drama`,
+    }))
+      .toEqual(Object.assign({}, initState, {
+        genre: `Drama`,
+      }));
+  });
+});
